h264: fix win32 compiler cast warning

+++ src/h264/nal.c
@@ -126,7 +126,7 @@ int h264_fu_hdr_decode(struct h264_fu *fu, struct mbuf *mb)
  */
 const uint8_t *h264_find_startcode(const uint8_t *p, const uint8_t *end)
 {
-	const uint8_t *a = p + 4 - ((long)p & 3);
+	const uint8_t *a = p + 4 - ((size_t)p & 3);
 
 	for (end -= 3; p < a && p < end; p++ ) {
 		if (p[0] == 0 && p[1] == 0 && p[2] == 1)
